Enable shorten-64-to-32 check

+++ cpp-ch/local-engine/Operator/BlockCoalesceOperator.cpp
@@ -5,7 +5,7 @@ namespace local_engine
 {
 void BlockCoalesceOperator::mergeBlock(DB::Block & block)
 {
-    block_buffer.add(block, 0, block.rows());
+    block_buffer.add(block, 0, static_cast<int>(block.rows()));
 }
 bool BlockCoalesceOperator::isFull()
 {
