+++ cpp-ch/local-engine/Parser/CHColumnToSparkRow.cpp
@@ -31,7 +31,7 @@ namespace local_engine
 {
 using namespace DB;
 
-int64_t calculateBitSetWidthInBytes(int32_t num_fields)
+int64_t calculateBitSetWidthInBytes(int64_t num_fields)
 {
     return ((num_fields + 63) / 64) * 8;
 }
@@ -48,7 +48,7 @@ int64_t roundNumberOfBytesToNearestWord(int64_t num_bytes)
 }
 
 
-void bitSet(char * bitmap, int32_t index)
+void bitSet(char * bitmap, size_t index)
 {
     int64_t mask = 1L << (index & 0x3f); // mod 64 and shift
     int64_t word_offset = (index >> 6) * 8;
@@ -58,7 +58,7 @@ void bitSet(char * bitmap, int32_t index)
     memcpy(bitmap + word_offset, &value, sizeof(int64_t));
 }
 
-ALWAYS_INLINE bool isBitSet(const char * bitmap, int32_t index)
+ALWAYS_INLINE bool isBitSet(const char * bitmap, size_t index)
 {
     assert(index >= 0);
     int64_t mask = 1 << (index & 63);
