+++ cpp-ch/local-engine/Shuffle/WriteBufferFromJavaOutputStream.cpp
@@ -14,7 +14,7 @@ void WriteBufferFromJavaOutputStream::nextImpl()
     size_t bytes_write = 0;
     while (offset() - bytes_write > 0)
     {
-        size_t copy_num = std::min(offset() - bytes_write, buffer_size);
+        jint copy_num = static_cast<jint>(std::min(offset() - bytes_write, buffer_size));
         env->SetByteArrayRegion(buffer, 0, copy_num, reinterpret_cast<const jbyte *>(this->working_buffer.begin() + bytes_write));
         safeCallVoidMethod(env, output_stream, output_stream_write, buffer, 0, copy_num);
         bytes_write += copy_num;
